# api/openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: Attesta Judicial Network API
  version: 0.1.0
  summary: Multi-court federation API for judicial log entries, artifacts, and cross-court verification.
  description: |
    The Judicial Network (JN) API is the HTTP surface for an Attesta
    deployment serving one or more courts on a single listener. The
    composer mounts three sub-surfaces:

      - `/v1/judicial/*` — domain-aware judicial flows (cases, appeals,
        enforcement, parties, onboarding, artifacts, verification,
        monitoring, consortium federation, delegation, topology).
      - `/v1/entries/*`, `/v1/artifacts/*`, `/v1/delegations/*`,
        `/v1/keys/*`, `/v1/dids`, `/v1/scope/*` — generic Attesta
        protocol surface (build/sign/submit, key custody, scope
        amendment).
      - `/v1/verify/*` — read-side verification (origin, authority,
        delegation, cross-log, fraud-proof).

    Routing is multi-tenant payload-driven: every entry's destination
    is sourced from `entry.Header.Destination`, looked up in the
    Bundle Registry. There is exactly one handler tree per route
    family, shared across all registered destinations.

    ### Build / sign / submit pattern

    Most write endpoints under `/v1/judicial/*` return a
    `BuildResponse` carrying the unsigned entry bytes and the
    signing payload. The caller signs the payload externally
    (typically through Privy or an SCW party signer) and submits the
...

    ### Authentication

...

...

...
  contact:
...

...
servers:
...
security:
  - mTLS: []
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────
paths:
...
  /v1/judicial/verification/cross-log-proof:
    post:
      tags: [Verification]
      summary: Verify a cross-log proof against source-court witness keys.
      operationId: verifyCrossLogProof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrossLogProofRequest'
      responses:
        '200':
          description: Verification verdict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrossLogProofResult'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ── Monitoring ────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    mTLS:
      type: mutualTLS
      description: |
        Client certificate carries the caller's DID in a SAN URI.
        Composer-level middleware lifts it into request context.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT validated against a configured JWKS endpoint. The `sub`
        claim is the caller DID.

  # ── Reusable parameters ───────────────────────────────────────────
...
  responses:
...
    BadRequest:
      description: Invalid request body or missing required fields.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid auth (no callerDID resolved).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    InternalError:
      description: Server-side error (typically a misconfigured dependency).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

  # ── Schemas ───────────────────────────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    CrossLogProofRequest:
      type: object
      required: [proof, source_log_did]
      properties:
        proof:
          type: object
          additionalProperties: true
          description: Marshaled types.CrossLogProof.
        source_log_did: { type: string }
        source_witness_keys_b64:
          type: array
          items: { type: string, format: byte }
        source_witness_quorum: { type: integer }
    CrossLogProofResult:
      type: object
      required: [verified]
      properties:
        verified: { type: boolean }
        error: { type: string }
...